requirements.txt
pandas
scikit-learn
torch
diffusers
Pillow
tf-keras
transformers>=4.40
accelerate>=0.30
sentencepiece


# genaivenv :v env
# set HF_TOKEN=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    :  token genai